requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
albumentations==1.4.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
pillow==9.3.0
timm==1.0.13
torch==2.5.1
torchvision
transformers==4.47.0
accelerate==1.6.0
pymupdf==1.26
# onnx export requires
onnxruntime-gpu==1.22.0
optimum[onnxruntime]>=1.16.0
accelerate
onnx==1.17.0
onnxconverter_common==1.14.0